+++ src/main/java/me/marin/lockout/lockout/goals/have_more/HaveMostDiamondBlocksGoal.java
@@ -1,15 +1,21 @@
 package me.marin.lockout.lockout.goals.have_more;
 
 import me.marin.lockout.Constants;
+import me.marin.lockout.LockoutTeam;
+import me.marin.lockout.LockoutTeamServer;
 import me.marin.lockout.lockout.Goal;
+import me.marin.lockout.lockout.interfaces.MostStatGoal;
 import me.marin.lockout.lockout.texture.CustomTextureRenderer;
+import me.marin.lockout.server.LockoutServer;
 import net.minecraft.client.gl.RenderPipelines;
 import net.minecraft.client.gui.DrawContext;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.Items;
 import net.minecraft.util.Identifier;
 
-public class HaveMostDiamondBlocksGoal extends Goal implements CustomTextureRenderer {
+import java.util.UUID;
+
+public class HaveMostDiamondBlocksGoal extends Goal implements CustomTextureRenderer, MostStatGoal {
 
     private static final ItemStack ITEM_STACK = Items.DIAMOND_BLOCK.getDefaultStack();
     
@@ -27,13 +33,22 @@ public class HaveMostDiamondBlocksGoal extends Goal implements CustomTextureRend
         return ITEM_STACK;
     }
 
-    private static final Identifier TEXTURE = Identifier.of(Constants.NAMESPACE, "textures/custom/up_arrow.png");
+    private static final Identifier TEXTURE = Identifier.of(Constants.NAMESPACE, "textures/custom/overlay/up_overlay.png");
     
     @Override
     public boolean renderTexture(DrawContext context, int x, int y, int tick) {
         context.drawItem(ITEM_STACK, x, y);
         context.drawTexture(RenderPipelines.GUI_TEXTURED, TEXTURE, x, y, 0,0, 16, 16, 16, 16);
         return true;
+    }
+
+    @Override
+    public int getStat(LockoutTeam team) {
+        int max = 0;
+        for (UUID uuid : ((LockoutTeamServer)team).getPlayers()) {
+            max = Math.max(max, LockoutServer.lockout.playerDiamondBlockCounts.getOrDefault(uuid, 0));
+        }
+        return max;
     }
 }