+++ src/main/java/me/marin/lockout/mixin/server/DecorationItemMixin.java
@@ -1,0 +1,49 @@
+package me.marin.lockout.mixin.server;
+
+import me.marin.lockout.Lockout;
+import me.marin.lockout.lockout.Goal;
+import me.marin.lockout.lockout.goals.misc.PlacePaintingGoal;
+import me.marin.lockout.server.LockoutServer;
+import net.minecraft.entity.EntityType;
+import net.minecraft.entity.decoration.AbstractDecorationEntity;
+import net.minecraft.entity.decoration.GlowItemFrameEntity;
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.item.DecorationItem;
+import net.minecraft.item.ItemStack;
+import net.minecraft.item.ItemUsageContext;
+import net.minecraft.util.ActionResult;
+import org.spongepowered.asm.mixin.Final;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+
+@Mixin(DecorationItem.class)
+public class DecorationItemMixin {
+
+    @Shadow @Final private EntityType<? extends AbstractDecorationEntity> entityType;
+
+    @Inject(method = "useOnBlock", at = @At("RETURN"))
+    public void onUseOnBlock(ItemUsageContext context, CallbackInfoReturnable<ActionResult> cir) {
+        if (context.getWorld().isClient()) return;
+        if (cir.getReturnValue() != ActionResult.SUCCESS) return;
+        if (this.entityType != EntityType.PAINTING) return;
+
+        Lockout lockout = LockoutServer.lockout;
+        if (!Lockout.isLockoutRunning(lockout)) return;
+
+        PlayerEntity player = context.getPlayer();
+
+        for (Goal goal : lockout.getBoard().getGoals()) {
+            if (goal == null) continue;
+            if (goal.isCompleted()) continue;
+
+            if (goal instanceof PlacePaintingGoal) {
+                lockout.completeGoal(goal, player);
+                return;
+            }
+        }
+    }
+
+}