+++ src/main/java/me/marin/lockout/mixin/server/FilledMapItemMixin.java
@@ -1,0 +1,44 @@
+package me.marin.lockout.mixin.server;
+
+import me.marin.lockout.Lockout;
+import me.marin.lockout.lockout.Goal;
+import me.marin.lockout.lockout.goals.misc.MapBannerWaypointGoal;
+import me.marin.lockout.server.LockoutServer;
+import net.minecraft.block.BlockState;
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.item.FilledMapItem;
+import net.minecraft.item.ItemUsageContext;
+import net.minecraft.registry.tag.BlockTags;
+import net.minecraft.util.ActionResult;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+
+@Mixin(FilledMapItem.class)
+public class FilledMapItemMixin {
+
+    @Inject(method = "useOnBlock", at = @At("RETURN"))
+    public void onUseOnBlock(ItemUsageContext context, CallbackInfoReturnable<ActionResult> cir) {
+        if (cir.getReturnValue() != ActionResult.SUCCESS) return;
+        PlayerEntity player = context.getPlayer();
+        if (player == null) return;
+        if (context.getWorld().isClient()) return;
+
+        Lockout lockout = LockoutServer.lockout;
+        if (!Lockout.isLockoutRunning(lockout)) return;
+
+        BlockState blockState = context.getWorld().getBlockState(context.getBlockPos());
+        if (blockState.isIn(BlockTags.BANNERS)) {
+            for (Goal goal : lockout.getBoard().getGoals()) {
+                if (goal == null) continue;
+                if (goal.isCompleted()) continue;
+
+                if (goal instanceof MapBannerWaypointGoal) {
+                    lockout.completeGoal(goal, player);
+                }
+            }
+        }
+    }
+
+}